refactor: extract pool label rendering into separate component
