+++ web/components/SankeyDiagram.tsx
@@ -12,6 +12,7 @@ import { useSankeyColors } from '@/hooks/useSankeyColors';
 import { getBranchColor } from '@/utils/sankeyColors';
 import { useTheme } from 'next-themes';
 import SankeyTooltip, { TooltipData } from './sankey/SankeyTooltip';
+import SankeyPoolLabels from './sankey/SankeyPoolLabels';
 
 interface SankeyDiagramProps {
   height: number;
@@ -428,194 +429,7 @@ export default function SankeyDiagram({
           setTooltipPosition(null);
         });
       
-      // Render pool labels next to the respective last merkle branch nodes
-      const renderPoolLabels = () => {
-        // Get the direct mapping from pool name to last branch name
-        const poolToLastBranchMap = sankeyDataProcessor.getLastMerkleBranchesForPools();
-        
-        if (poolToLastBranchMap.size === 0) {
-          console.warn('No pool to branch mappings found');
-          return null;
-        }
-        
-        // Log all pools we're trying to render
-        const allPoolNames = Array.from(poolToLastBranchMap.keys());
-        console.log('Attempting to render labels for pools:', allPoolNames);
-        
-        // Create mapping from branch name to node index for quick lookups
-        const branchNameToNodeIndex = new Map<string, number>();
-        nodes.forEach((node: any, index: number) => {
-          if (node.type === 'branch') {
-            branchNameToNodeIndex.set(node.name, index);
-            // Also add lowercase version for case-insensitive matching
-            branchNameToNodeIndex.set(node.name.toLowerCase(), index);
-          }
-        });
-        
-        // Calculate global max X for fallback alignment
-        const globalMaxX = Math.max(...nodes.map((n: any) => n.x1 || 0)) + 5;
-        
-        // For vertical spacing management
-        const yPositionsByColumn: {[colX: number]: number[]} = {};
-        
-        // Create an array of pool label elements
-        const poolLabels: React.ReactNode[] = [];
-        let poolLabelCount = 0;
-        
-        // Process each pool
-        poolToLastBranchMap.forEach((branchName, poolName) => {
-          // Find the node index for this branch
-          let nodeIndex = branchNameToNodeIndex.get(branchName);
-          
-          // Try case-insensitive match if needed
-          if (nodeIndex === undefined) {
-            nodeIndex = branchNameToNodeIndex.get(branchName.toLowerCase());
-          }
-          
-          // Skip if we can't find the node
-          if (nodeIndex === undefined) {
-            console.warn(`Could not find node for branch ${branchName} (pool ${poolName})`);
-            return;
-          }
-          
-          const node = nodes[nodeIndex];
-          if (!node) {
-            console.warn(`Node index ${nodeIndex} not found for pool ${poolName}`);
-            return;
-          }
-          
-          // Calculate position for label - always place next to its actual branch node
-          const nodeX = Math.round(node.x1 || 0);
-          const labelX = nodeX + 10; // Fixed offset from node
-          
-          // Initialize the column's y-positions array if it doesn't exist
-          if (!yPositionsByColumn[nodeX]) {
-            yPositionsByColumn[nodeX] = [];
-          }
-          
-          // Find the node's center y position
-          const nodeCenterY = (node.y0 || 0) + ((node.y1 || 0) - (node.y0 || 0)) / 2;
-          
-          // Find a suitable y position that doesn't overlap with other labels
-          let labelY = nodeCenterY;
-          const minDistance = 25; // Minimum vertical distance between labels
-          
-          // Check if this position overlaps with any existing label in this column
-          while (yPositionsByColumn[nodeX].some(y => Math.abs(y - labelY) < minDistance)) {
-            labelY += minDistance; // Move label down if there would be overlap
-          }
-          
-          // Store this position
-          yPositionsByColumn[nodeX].push(labelY);
-          
-          console.log(`Adding pool label: ${poolName} at position (${labelX},${labelY}) for branch ${branchName} (node index: ${nodeIndex})`);
-          
-          // Create the label
-          poolLabels.push(
-            <g key={`pool-label-${poolName}`} className="pool-label">
-              {/* Line connecting label to node */}
-              <line
-                x1={(node.x1 || 0) + 2}
-                y1={nodeCenterY}
-                x2={labelX - 4}
-                y2={labelY}
-                stroke={theme === 'dark' ? '#aaa' : '#666'}
-                strokeWidth={1.5}
-                strokeDasharray="3,2"
-              />
-              
-              {/* Background rectangle for better readability */}
-              <rect
-                x={labelX - 4}
-                y={labelY - 10}
-                width={poolName.length * 8 + 10}
-                height={20}
-                rx={4}
-                fill={theme === 'dark' ? '#333' : '#f0f0f0'}
-                stroke={colors.poolLabel}
-                strokeWidth={1}
-              />
-              
-              {/* Pool name text */}
-              <text
-                x={labelX}
-                y={labelY + 4}
-                fontWeight="bold"
-                fill={colors.poolLabel}
-                fontSize="12"
-                style={{
-                  paintOrder: 'stroke',
-                  stroke: theme === 'dark' ? '#222' : '#fff',
-                  strokeWidth: '2px',
-                  strokeLinecap: 'round',
-                  strokeLinejoin: 'round',
-                }}
-              >
-                {poolName}
-              </text>
-            </g>
-          );
-          poolLabelCount++;
-        });
-        
-        console.log(`Rendered ${poolLabelCount} pool labels from ${poolToLastBranchMap.size} pools`);
-        
-        // If no pool labels were found via the usual method, try a direct approach as fallback
-        if (poolLabelCount === 0) {
-          console.warn('No pool labels were created via node mapping, using direct pool data');
-          
-          // Get all pools directly from the data processor
-          const poolToBranch = sankeyDataProcessor.getLastMerkleBranchesForPools();
-          const rightEdge = globalMaxX; // Use the previously defined max X
-          
-          // Render labels directly using the pool information
-          let verticalOffset = 50; // Start with an offset from top
-          poolToBranch.forEach((branchName, poolName) => {
-            console.log(`Direct pool label: ${poolName} (branch: ${branchName})`);
-            
-            poolLabels.push(
-              <g key={`direct-pool-label-${poolName}`} className="pool-label direct">
-                {/* Background rectangle */}
-                <rect
-                  x={rightEdge - 4}
-                  y={verticalOffset - 10}
-                  width={poolName.length * 8 + 10}
-                  height={20}
-                  rx={4}
-                  fill={theme === 'dark' ? '#333' : '#f0f0f0'}
-                  stroke={colors.poolLabel}
-                  strokeWidth={1}
-                />
-                
-                {/* Pool name text */}
-                <text
-                  x={rightEdge}
-                  y={verticalOffset + 4}
-                  fontWeight="bold"
-                  fill={colors.poolLabel}
-                  fontSize="12"
-                  style={{
-                    paintOrder: 'stroke',
-                    stroke: theme === 'dark' ? '#222' : '#fff',
-                    strokeWidth: '2px',
-                    strokeLinecap: 'round',
-                    strokeLinejoin: 'round',
-                  }}
-                >
-                  {poolName}
-                </text>
-              </g>
-            );
-            
-            verticalOffset += 25; // Space labels vertically
-            poolLabelCount++;
-          });
-          
-          console.log(`Added ${poolLabelCount} direct pool labels as fallback`);
-        }
-        
-        return <>{poolLabels}</>;
-      };
+
       
       // Add node labels if enabled
       if (showLabels) {
@@ -673,219 +487,47 @@ export default function SankeyDiagram({
         });
       }
       
-      // Find the absolute rightmost branch node position for alignment
-      const branchNodes = nodes.filter((n: any) => n.type === 'branch');
-      const rightmostBranchX1 = Math.max(...branchNodes.map((n: any) => n.x1 || 0));
-      
-      // Create a map from branch name to node index
-      const branchNameToNode = new Map();
-      branchNodes.forEach((node: any) => {
-        branchNameToNode.set(node.name, node);
-      });
-      
-      // Group branch nodes by their horizontal position to identify columns
-      const branchNodesByX = new Map<number, any[]>();
-      branchNodes.forEach((node: any) => {
-        const x = Math.round(node.x1 || 0);
-        if (!branchNodesByX.has(x)) {
-          branchNodesByX.set(x, []);
-        }
-        branchNodesByX.get(x)!.push(node);
-      });
-      
-      // Interface for pool label info collection
-      interface PoolLabelInfo {
-        nodeX: number;
-        nodeY: number;
-        poolName: string;
-        branchName: string;
+      // Render pool labels using the extracted SankeyPoolLabels component
+      // The component uses D3 for direct SVG manipulation, so we need to trigger it here
+      if (svg && nodes && sankeyDataProcessor) {
+        // Create a temporary container for the React component that will handle D3 rendering
+        const poolLabelsContainer = document.createElement('div');
+        poolLabelsContainer.style.display = 'none'; // Hidden since it's just for triggering rendering
+        document.body.appendChild(poolLabelsContainer);
+        
+        // Use React.createElement to instantiate SankeyPoolLabels
+        // This will trigger the useEffect inside the component to render pool labels via D3
+        const poolLabelsElement = React.createElement(SankeyPoolLabels, {
+          nodes,
+          sankeyDataProcessor,
+          width,
+          height,
+          theme: theme || 'light',
+          colors: {
+            poolLabel: colors.poolLabel,
+            ...colors as any  // Cast to any to satisfy index signature
+          },
+          svg
+        });
+        
+        // Render the component to trigger its D3 rendering logic
+        import('react-dom/client').then(({ createRoot }) => {
+          const root = createRoot(poolLabelsContainer);
+          root.render(poolLabelsElement);
+          
+          // Clean up after a short delay to allow rendering to complete
+          setTimeout(() => {
+            root.unmount();
+            document.body.removeChild(poolLabelsContainer);
+          }, 100);
+        }).catch(err => {
+          console.error('Error loading React DOM:', err);
+          // Fallback: Clean up container
+          if (document.body.contains(poolLabelsContainer)) {
+            document.body.removeChild(poolLabelsContainer);
+          }
+        });
       }
-      
-      const poolLabelsToAdd: PoolLabelInfo[] = [];
-      
-      // Get the direct mapping from pool name to last branch name
-      const poolToLastBranchMap = sankeyDataProcessor.getLastMerkleBranchesForPools();
-      console.log("All pools with their last branches:", Object.fromEntries(poolToLastBranchMap));
-      
-      // Group pools by their last merkle branch
-      const branchToPoolsMap = new Map<string, string[]>();
-      
-      poolToLastBranchMap.forEach((branchName, poolName) => {
-        if (!branchToPoolsMap.has(branchName)) {
-          branchToPoolsMap.set(branchName, []);
-        }
-        branchToPoolsMap.get(branchName)!.push(poolName);
-      });
-      
-      // Sort pools alphabetically within each branch group
-      branchToPoolsMap.forEach((pools, branch) => {
-        pools.sort(); // Sort alphabetically
-      });
-      
-      // Process each branch with its associated pools
-      branchToPoolsMap.forEach((poolNames, branchName) => {
-        // Find the node for this branch name
-        let branchNode = undefined;
-        
-        // Try exact match first
-        for (const node of branchNodes) {
-          if (node.name === branchName) {
-            branchNode = node;
-            break;
-          }
-        }
-        
-        // Try case-insensitive match if needed
-        if (!branchNode) {
-          const lowerBranchName = branchName.toLowerCase();
-          for (const node of branchNodes) {
-            if (node.name.toLowerCase() === lowerBranchName) {
-              branchNode = node;
-              break;
-            }
-          }
-        }
-        
-        // Skip if we can't find the node
-        if (!branchNode) {
-          console.warn(`Could not find node for branch ${branchName} (pools: ${poolNames.join(', ')})`);
-          return;
-        }
-        
-        // Get the position for this node
-        const nodeX = branchNode.x1; // right edge
-        const nodeY = branchNode.y0 + ((branchNode.y1 - branchNode.y0) / 2); // vertical center
-        
-        console.log(`Found node for branch ${branchName} at position (${nodeX}, ${nodeY}) with pools: ${poolNames.join(', ')}`);
-        
-        // Add to our labels list
-        poolLabelsToAdd.push({
-          nodeX,
-          nodeY,
-          poolName: poolNames.join('\n'), // We'll split this later when rendering
-          branchName
-        });
-      });
-      
-      // Sort labels by vertical position
-      poolLabelsToAdd.sort((a, b) => a.nodeY - b.nodeY);
-      
-      // Remove any existing pool labels first
-      svg.selectAll(".pool-labels-group").remove();
-      
-      // Create a new pool label group
-      const poolLabelGroup = svg.append("g")
-        .attr("class", "pool-labels-group");
-      
-      // Calculate vertical spacing for labels (if multiple pools share the same x-coordinate)
-      const LINE_HEIGHT = 16; // Height per line of text
-      const LABEL_PADDING = 6; // Padding inside label box
-      const LABEL_OFFSET_X = 20; // Horizontal offset from the node
-      
-      // Add pool labels with nice styling and connecting lines
-      poolLabelsToAdd.forEach((labelInfo, index) => {
-        // Calculate horizontal position to the right of the node
-        const x = labelInfo.nodeX + LABEL_OFFSET_X;
-        
-        // Base vertical position at node center
-        let y = labelInfo.nodeY;
-        
-        // Split the pool names (they're joined with newlines)
-        const poolNames = labelInfo.poolName.split('\n');
-        
-        // Calculate background rectangle dimensions
-        const maxLabelLength = Math.max(...poolNames.map(name => name.length));
-        const TEXT_PADDING_HORIZONTAL = 4; // Equal horizontal padding for both sides
-        const rectWidth = maxLabelLength * 7.2 + (TEXT_PADDING_HORIZONTAL * 2); // Equal padding on both sides
-        
-        // More precise height calculation to ensure equal padding top and bottom
-        // BASE_TEXT_SIZE is the actual height of the font (not the same as LINE_HEIGHT)
-        const BASE_TEXT_SIZE = 12; // Font size in pixels
-        // LINE_HEIGHT is space between lines
-        
-        // Calculate total content height needed
-        const textContentHeight = poolNames.length * LINE_HEIGHT - (LINE_HEIGHT - BASE_TEXT_SIZE);
-        
-        // Add the same padding to top and bottom
-        const rectHeight = textContentHeight + (LABEL_PADDING * 2);
-        
-        // Ensure the label stays within the visible height area
-        if (y + (rectHeight / 2) > height) {
-          // If label would extend below the bottom, move it up
-          y = height - (rectHeight / 2) - 10; // 10px buffer from bottom edge
-        }
-        if (y - (rectHeight / 2) < 0) {
-          // If label would extend above the top, move it down
-          y = (rectHeight / 2) + 10; // 10px buffer from top edge
-        }
-        
-        // Use a completely different approach with explicit positioning
-        // First figure out how much total vertical space we need
-        const FONT_SIZE = 12;
-        const TEXT_PADDING = 1; // Minimal padding between text and box edge
-        
-        // Create the group element for this label
-        const labelGroup = poolLabelGroup.append("g")
-          .attr("class", "pool-label")
-          .attr("transform", `translate(${x}, ${y})`);
-        
-        // Add connecting line
-        labelGroup.append("line")
-          .attr("x1", -LABEL_OFFSET_X) // Connect from the node edge
-          .attr("y1", 0) // Center of node
-          .attr("x2", -5) // Small offset before the label
-          .attr("y2", 0)
-          .attr("stroke", theme === 'dark' ? '#aaa' : '#666')
-          .attr("stroke-width", 1.5)
-          .attr("stroke-dasharray", "3,2");
-          
-        // This will be used to build our label from scratch
-        // with precise control over padding
-        
-        // Step 1: Calculate total content height without extra spacing
-        const contentHeight = poolNames.length * LINE_HEIGHT;
-        const halfContentHeight = contentHeight / 2;
-        
-        // Step 2: Set the vertical boundaries of our content
-        const topEdge = -halfContentHeight;
-        const bottomEdge = halfContentHeight;
-        
-        // Step 3: Create the background rectangle with equal padding
-        labelGroup.append("rect")
-          .attr("x", -5)
-          .attr("y", topEdge - TEXT_PADDING) // Top of content minus padding
-          .attr("width", rectWidth)
-          .attr("height", contentHeight + (TEXT_PADDING * 2)) // Content height plus padding top and bottom
-          .attr("rx", 4)
-          .attr("fill", theme === 'dark' ? '#333' : '#f0f0f0')
-          .attr("stroke", colors.poolLabel)
-          .attr("stroke-width", 1);
-        
-        // Step 4: Position each line of text
-        poolNames.forEach((name, i) => {
-          // Calculate center position for each text line
-          // First line starts at the top edge of content area
-          // Each subsequent line is positioned LINE_HEIGHT below the previous
-          const textY = topEdge + (i * LINE_HEIGHT) + (LINE_HEIGHT / 2);
-          
-          // Add the text directly (not using tspan)
-          labelGroup.append("text")
-            .attr("x", TEXT_PADDING_HORIZONTAL - 5) // Account for rect x offset with new padding
-            .attr("y", textY) // Center of this text line
-            .attr("text-anchor", "start")
-            .attr("dominant-baseline", "middle") // Center text vertically on its position
-            .attr("fill", colors.poolLabel)
-            .attr("stroke", theme === 'dark' ? '#222' : '#fff')
-            .attr("stroke-width", 0.5)
-            .attr("paint-order", "stroke")
-            .attr("font-size", `${FONT_SIZE}px`)
-            .attr("font-weight", "600")
-            .text(name);
-        });
-      });
-      
-      // Log for debugging
-      console.log(`Added ${poolLabelsToAdd.length} pool labels to the diagram`);
       
     } catch (err) {
       console.error("Error rendering diagram:", err);
